split pdf model client and server
