+++ commandline.js
@@ -1,4 +1,4 @@
-import { pdfModelServer } from './src/pdf/pdf-model.js'
+import { pdfModelServer } from './src/pdf/pdf-model-server.js'
 import { testPlanCommandLine } from './src/command-line/test-plan-commandline.js'
 import { getPDFContentForServer } from './src/pdf/models/template.js'
