+++ example-pdf-creation-client.js
@@ -1,4 +1,4 @@
-import { pdfModelClient } from './src/pdf/pdf-model.js'
+import { pdfModelClient } from './src/pdf/pdf-model-client.js'
 import { getPDFContentForClient } from './src/pdf/models/template.js'
 
 let pdfContent = getPDFContentForClient('My project', 'frontend', 'chrome')
