+++ example-pdf-creation-server.js
@@ -1,4 +1,4 @@
-import { pdfModelServer } from './src/pdf/pdf-model.js'
+import { pdfModelServer } from './src/pdf/pdf-model-server.js'
 import { getPDFContentForServer } from './src/pdf/models/template.js'
 
 let pdfContent = getPDFContentForServer('My project', 'frontend', 'chrome')
