+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "test-artifacts",
-  "version": "1.0.4",
+  "version": "1.0.5",
   "description": "",
   "type": "module",
   "main": "index.js",
