+++ src/pdf/pdf-model-server.js
@@ -1,6 +1,4 @@
 import PdfPrinter from 'pdfmake';
-import pdfMake from 'pdfmake/build/pdfmake.js';
-import pdfFonts from 'pdfmake/build/vfs_fonts.js';
 import fs from 'fs';
 
 let fonts = {
@@ -26,14 +24,4 @@ export async function pdfModelServer(content){
     writeStream.on('error', (err) => {
         console.error('Error generating PDF:', err);
     });
-}
-
-export async function pdfModelClient(content, fileName){
-
-    try{
-      pdfMake.vfs = pdfFonts.pdfMake.vfs; // Setting the virtual file system for fonts
-      pdfMake.createPdf(content).download(`${fileName}.pdf`);
-    } catch (error) {
-      console.error('Error generating PDF:', error);
-    }
 }
